+++ src/App/Fossa/Config/ConfigFile.hs
@@ -206,6 +206,7 @@ data ConfigFile = ConfigFile
   , configCustomLicenseSearch :: Maybe [ConfigGrepEntry]
   , configKeywordSearch :: Maybe [ConfigGrepEntry]
   , configReachability :: Maybe ReachabilityConfigFile
+  , configEnableGoBinaryAnalysis :: Maybe Bool
   , configOrgWideCustomLicenseConfigPolicy :: OrgWideCustomLicenseConfigPolicy
   , configConfigFilePath :: Path Abs File
   }
@@ -309,6 +310,7 @@ instance FromJSON (Path Abs File -> ConfigFile) where
       <*> obj .:? "customLicenseSearch"
       <*> obj .:? "experimentalKeywordSearch"
       <*> obj .:? "reachability"
+      <*> obj .:? "enableGoBinaryAnalysis"
       <*> parseIgnoreOrgWideCustomLicenseScanConfigs obj
     where
       parseIgnoreOrgWideCustomLicenseScanConfigs obj = do