+++ test/App/Fossa/Config/AnalyzeSpec.hs
@@ -58,6 +58,7 @@ configFileWithTargets only exclude excludeManifestStrategies =
     , configCustomLicenseSearch = Nothing
     , configKeywordSearch = Nothing
     , configReachability = Nothing
+    , configEnableGoBinaryAnalysis = Nothing
     , configOrgWideCustomLicenseConfigPolicy = Use
     , configConfigFilePath = configPath
     }