+++ test/App/Fossa/Config/ReleaseGroup/CreateSpec.hs
@@ -90,6 +90,7 @@ configFile path =
     , configCustomLicenseSearch = Nothing
     , configKeywordSearch = Nothing
     , configReachability = Nothing
+    , configEnableGoBinaryAnalysis = Nothing
     , configOrgWideCustomLicenseConfigPolicy = Use
     , configConfigFilePath = path
     }