+++ test/App/Fossa/Configuration/TelemetryConfigSpec.hs
@@ -85,6 +85,7 @@ defaultConfigFile =
     , configOrgWideCustomLicenseConfigPolicy = Use
     , configConfigFilePath = configPath
     , configReachability = Nothing
+    , configEnableGoBinaryAnalysis = Nothing
     }
 
 mockApiKeyRaw :: Text